backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.1
langchain==0.3.0
langchain-community==0.3.0
langchain-google-genai==2.0.0
langchain-text-splitters==0.3.0
faiss-cpu==1.7.4
websockets==12.0
httpx==0.26.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1